add function to check if user exists

+++ actions/login.php
@@ -8,16 +8,17 @@ require_once '../includes/functions.php';
 $email = filter_input(INPUT_POST, 'email', FILTER_VALIDATE_EMAIL, FILTER_SANITIZE_EMAIL);
 $password = filter_input(INPUT_POST, 'password', FILTER_SANITIZE_SPECIAL_CHARS);
 
-try {
-    $user = getOneUser($email);
+if (!userExists($email)) {
+    echo 'User não encontrado';
+    exit;
+}
 
-    if (password_verify($password, $user->password)) {
-        $_SESSION['user_id'] = $user->user_id;
-        redirect('../../views/home.php');
-    } else {
-        echo '<p>Senha ou usuario incorretos</p><br>';
-        echo "<a href='../views/login.php'>Voltar</a>";
-    }
-} catch (Exception $e) {
-    echo $e->getMessage();
+$user = getOneUser($email);
+
+if (password_verify($password, $user->password)) {
+    $_SESSION['user_id'] = $user->user_id;
+    redirect('../../views/home.php');
+} else {
+    echo '<p>Senha ou usuario incorretos</p><br>';
+    echo "<a href='../views/login.php'>Voltar</a>";
 }
